add get series use case

+++ app/src/main/java/com/elhady/movies/data/remote/response/RatedSeriesDto.kt
@@ -28,7 +28,7 @@ data class RatedSeriesDto(
     @SerializedName("poster_path")
     val posterPath: String?,
     @SerializedName("rating")
-    val rating: Double?,
+    val rating: Float?,
     @SerializedName("vote_average")
     val voteAverage: Double?,
     @SerializedName("vote_count")
